test: split Development panel timing tests under economy 400-line gate (#4175)
Extract shared fixture and reuse guards into sibling files so
development_panel_open_path_timing_test.dart stays within the
colonizethis_economy test file-size ceiling.
